requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.3.0
audioread==3.0.1
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
decorator==5.1.1
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
flatbuffers==24.12.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.4.2
keras==3.8.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.1.0
mysqlclient==2.2.6
namex==0.0.8
numba==0.60.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==24.3.0
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
soundfile==0.13.0
soxr==0.5.0.post1
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
zope.interface==7.2
